Revert "Update Config "
