+++ exaspim_control/experimental/gui_config.yaml
@@ -91,7 +91,7 @@ acquisition_view:
   fov_dimensions: [ 10.615616, 7.95872 ]
   coordinate_plane: [ '-x', 'y', 'z' ]
   unit: 'mm'
-  properties:
+  settings:
     cameras: [ binning ]
     lasers: [ power_setpoint_mw ]
     focusing_stages: [ position_mm ]
