+++ exaspim_control/simulated/gui_config.yaml
@@ -77,7 +77,7 @@ acquisition_view:
   fov_dimensions: [ 10.672384, 8.00128 ]
   coordinate_plane: [ '-x', 'y', 'z' ]
   unit: 'mm'
-  properties:
+  settings:
     cameras: [ binning ]
     lasers: [ power_setpoint_mw ]
     focusing_stages: [position_mm]
