+++ exaspim_control/test/camera_only/acquisition2.yaml
@@ -15,7 +15,7 @@ acquisition:
         module: Writer
         init:
           path: D:\\local_test
-        properties:
+        settings:
           compression: lz4shuffle
           data_type: uint16
       robocopy:
@@ -31,7 +31,7 @@ acquisition:
         module: MaxProjection
         init:
           path: D:\\local_test
-        properties:
+        settings:
           data_type: uint16
           projection_count_px: 233
   tiles:
